Tests/LibWeb/Text/expected/css/animated-background-image-pauses-for-inactive-document.txt

timer stopped after iframe removal: true
animation restarted after iframe reattach: true
